# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/admin/community-sounds": {
      "get": {
        "operationId": "AdminCommunityController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": ["pending", "approved", "rejected"],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminUserSoundDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Moderasyon listesi (durum filtreli)",
        "tags": ["admin-community"]
      }
    },
    "/v1/admin/community-sounds/{id}/file": {
      "get": {
        "operationId": "AdminCommunityController_file",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string"
                    },
                    "expiresIn": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Moderatör önizlemesi: kısa ömürlü dinleme URL’i (10 dk)",
        "tags": ["admin-community"]
      }
    },
    "/v1/admin/community-sounds/{id}/moderate": {
      "post": {
        "operationId": "AdminCommunityController_decide",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModerateSoundDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserSoundDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Onayla / reddet (redde gerekçe zorunlu)",
        "tags": ["admin-community"]
      }
    }
  },
  "info": {
    "title": "NOCTA API",
    "description": "Sleep Identity Platform — v1",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AdminUserSoundDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "byteSize": {
            "type": "object",
            "nullable": true
          },
          "durationSeconds": {
            "type": "number"
          },
          "rejectionReason": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "moderatedAt": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "byteSize",
          "durationSeconds",
          "rejectionReason",
          "createdAt",
          "userId",
          "moderatedAt"
        ]
      },
      "ModerateSoundDto": {
        "type": "object",
        "properties": {
          "decision": {
            "type": "string",
            "enum": ["approve", "reject"]
          },
          "rejectionReason": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": ["decision"]
      }
    }
  }
}